Refactor: change all names with 'stackoverflow' to 'waffleoverflow'
