--- src/main/kotlin/com/wafflestudio/stackoverflow/StackoverflowApplication.kt
@@ -1,11 +1,0 @@
-package com.wafflestudio.stackoverflow
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class StackoverflowApplication
-
-fun main(args: Array<String>) {
-    runApplication<StackoverflowApplication>(*args)
-}
