--- src/main/kotlin/com/wafflestudio/stackoverflow/domain/user/exception/UserAlreadyExistsException.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.stackoverflow.domain.user.exception
-
-import com.wafflestudio.stackoverflow.global.common.exception.ConflictException
-import com.wafflestudio.stackoverflow.global.common.exception.ErrorType
-
-class UserAlreadyExistsException(detail: String = "") :
-    ConflictException(ErrorType.USER_ALREADY_EXISTS, detail)
