--- src/main/kotlin/com/wafflestudio/stackoverflow/global/common/exception/ErrorResponse.kt
@@ -1,11 +1,0 @@
-package com.wafflestudio.stackoverflow.global.common.exception
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class ErrorResponse(
-    @JsonProperty("error_code")
-    val errorCode: Int,
-    @JsonProperty("error_message")
-    val errorMessage: String = "",
-    val detail: String = ""
-)
