--- src/main/kotlin/com/wafflestudio/stackoverflow/global/config/AppConfig.kt
@@ -1,6 +1,0 @@
-package com.wafflestudio.stackoverflow.global.config
-
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class AppConfig
