+++ src/main/kotlin/com/wafflestudio/waffleoverflow/WaffleoverflowApplication.kt
@@ -1,0 +1,11 @@
+package com.wafflestudio.waffleoverflow
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class WaffleoverflowApplication
+
+fun main(args: Array<String>) {
+    runApplication<WaffleoverflowApplication>(*args)
+}
