+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/model/BaseEntity.kt
@@ -1,0 +1,13 @@
+package com.wafflestudio.waffleoverflow.domain.model
+
+import javax.persistence.GeneratedValue
+import javax.persistence.GenerationType
+import javax.persistence.Id
+import javax.persistence.MappedSuperclass
+
+@MappedSuperclass
+open class BaseEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    open val id: Long = 0
+}
