+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/user/api/UserController.kt
@@ -1,0 +1,23 @@
+package com.wafflestudio.waffleoverflow.domain.user.api
+
+import com.wafflestudio.waffleoverflow.domain.user.dto.UserDto
+import com.wafflestudio.waffleoverflow.domain.user.service.UserService
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+import javax.validation.Valid
+
+@RestController
+@RequestMapping("/api/v1/user")
+class UserController(
+    private val userService: UserService,
+) {
+    @PostMapping("/")
+    fun signup(@Valid @RequestBody signupRequest: UserDto.SignupRequest): ResponseEntity<UserDto.Response> {
+        val user = userService.signup(signupRequest)
+        return ResponseEntity(UserDto.Response(user), HttpStatus.CREATED)
+    }
+}
