+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/user/dto/UserDto.kt
@@ -1,0 +1,29 @@
+package com.wafflestudio.waffleoverflow.domain.user.dto
+
+import com.wafflestudio.waffleoverflow.domain.user.model.User
+import javax.validation.constraints.NotBlank
+
+class UserDto {
+    data class Response(
+        val id: Long,
+        val email: String,
+        val name: String,
+    ) {
+        constructor(user: User) : this(
+            id = user.id,
+            email = user.email,
+            name = user.name,
+        )
+    }
+
+    data class SignupRequest(
+        @field:NotBlank
+        val email: String,
+
+        @field:NotBlank
+        val name: String,
+
+        @field:NotBlank
+        val password: String,
+    )
+}
