+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/user/model/User.kt
@@ -1,0 +1,21 @@
+package com.wafflestudio.waffleoverflow.domain.user.model
+
+import com.wafflestudio.waffleoverflow.domain.model.BaseEntity
+import javax.persistence.Column
+import javax.persistence.Entity
+import javax.persistence.Table
+import javax.validation.constraints.NotBlank
+
+@Entity
+@Table(name = "user")
+class User(
+    @Column(unique = true)
+    @field:NotBlank
+    val email: String,
+
+    @field:NotBlank
+    var name: String,
+
+    @field:NotBlank
+    var password: String,
+) : BaseEntity()
