+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/user/repository/UserRepository.kt
@@ -1,0 +1,8 @@
+package com.wafflestudio.waffleoverflow.domain.user.repository
+
+import com.wafflestudio.waffleoverflow.domain.user.model.User
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository : JpaRepository<User, Long?> {
+    fun existsByEmail(email: String): Boolean
+}
