+++ src/main/kotlin/com/wafflestudio/waffleoverflow/domain/user/service/UserService.kt
@@ -1,0 +1,21 @@
+package com.wafflestudio.waffleoverflow.domain.user.service
+
+import com.wafflestudio.waffleoverflow.domain.user.dto.UserDto
+import com.wafflestudio.waffleoverflow.domain.user.exception.UserAlreadyExistsException
+import com.wafflestudio.waffleoverflow.domain.user.model.User
+import com.wafflestudio.waffleoverflow.domain.user.repository.UserRepository
+import org.springframework.stereotype.Service
+
+@Service
+class UserService(
+    private val userRepository: UserRepository,
+) {
+    fun signup(signupRequest: UserDto.SignupRequest): User {
+        if (userRepository.existsByEmail(signupRequest.email)) throw UserAlreadyExistsException()
+        val email = signupRequest.email
+        val name = signupRequest.name
+        val password = signupRequest.password
+
+        return userRepository.save(User(email, name, password))
+    }
+}
