+++ src/main/kotlin/com/wafflestudio/waffleoverflow/global/common/exception/ErrorResponse.kt
@@ -1,0 +1,11 @@
+package com.wafflestudio.waffleoverflow.global.common.exception
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+data class ErrorResponse(
+    @JsonProperty("error_code")
+    val errorCode: Int,
+    @JsonProperty("error_message")
+    val errorMessage: String = "",
+    val detail: String = ""
+)
