+++ src/main/kotlin/com/wafflestudio/waffleoverflow/global/common/exception/ErrorType.kt
@@ -1,0 +1,16 @@
+package com.wafflestudio.waffleoverflow.global.common.exception
+
+enum class ErrorType(
+    val code: Int
+) {
+    INVALID_REQUEST(0),
+
+    NOT_ALLOWED(3000),
+
+    DATA_NOT_FOUND(4000),
+
+    CONFLICT(9000),
+    USER_ALREADY_EXISTS(9001),
+
+    SERVER_ERROR(10000)
+}
