+++ src/main/kotlin/com/wafflestudio/waffleoverflow/global/common/exception/ExceptionControllerAdvice.kt
@@ -1,0 +1,27 @@
+package com.wafflestudio.waffleoverflow.global.common.exception
+
+import org.slf4j.LoggerFactory
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice
+class ExceptionControllerAdvice() {
+    private val logger = LoggerFactory.getLogger(this.javaClass.name)
+    @ExceptionHandler(value = [DataNotFoundException::class])
+    fun notfound(e: WaffleOverFlowException) =
+        ResponseEntity(ErrorResponse(e.errorType.code, e.errorType.name, e.detail), HttpStatus.NOT_FOUND)
+
+    @ExceptionHandler(value = [InvalidRequestException::class])
+    fun badRequest(e: WaffleOverFlowException) =
+        ResponseEntity(ErrorResponse(e.errorType.code, e.errorType.name, e.detail), HttpStatus.BAD_REQUEST)
+
+    @ExceptionHandler(value = [NotAllowedException::class])
+    fun notAllowed(e: WaffleOverFlowException) =
+        ResponseEntity(ErrorResponse(e.errorType.code, e.errorType.name, e.detail), HttpStatus.FORBIDDEN)
+
+    @ExceptionHandler(value = [ConflictException::class])
+    fun conflict(e: WaffleOverFlowException) =
+        ResponseEntity(ErrorResponse(e.errorType.code, e.errorType.name, e.detail), HttpStatus.CONFLICT)
+}
