+++ src/main/kotlin/com/wafflestudio/waffleoverflow/global/config/AppConfig.kt
@@ -1,0 +1,6 @@
+package com.wafflestudio.waffleoverflow.global.config
+
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+class AppConfig
