+++ src/main/kotlin/com/wafflestudio/waffleoverflow/global/config/SecurityConfig.kt
@@ -1,0 +1,15 @@
+package com.wafflestudio.waffleoverflow.global.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity
+import org.springframework.security.config.annotation.web.builders.HttpSecurity
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter
+
+@Configuration
+@EnableGlobalMethodSecurity(prePostEnabled = true)
+@EnableWebSecurity
+class SecurityConfig() : WebSecurityConfigurerAdapter() {
+    override fun configure(http: HttpSecurity) {
+    }
+}
