--- src/test/kotlin/com/wafflestudio/stackoverflow/StackoverflowApplicationTests.kt
@@ -1,12 +1,0 @@
-package com.wafflestudio.stackoverflow
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class StackoverflowApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-}
