+++ src/test/kotlin/com/wafflestudio/waffleoverflow/WaffleoverflowApplicationTests.kt
@@ -1,0 +1,12 @@
+package com.wafflestudio.waffleoverflow
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class WaffleoverflowApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+}
